Fixing metric names
